chore: 重构线图层,更新例子
