--- dev-demos/features/line/arc3d_linear.md
@@ -1,2 +1,0 @@
-### arc3d
-<code src="./demos/arc_linear.tsx"></code>
