+++ dev-demos/features/line/demos/line.tsx
@@ -16,7 +16,7 @@ export default () => {
   useEffect(() => {
     const scene = new Scene({
       id: 'map',
-    renderer: process.env.renderer,
+      renderer: process.env.renderer,
       map: new GaodeMap({
         center: [105, 32],
         zoom: 4,
@@ -75,9 +75,7 @@ export default () => {
       .shape('line')
       .color('#f00')
       .style({
-        opacity:1,
-        sourceColor: '#f00',
-        targetColor: '#0f0',
+        opacity:0.6
       });
 
     scene.on('loaded', () => {
