+++ dev-demos/features/line/demos/lineArc3d.tsx
@@ -13,7 +13,7 @@ export default () => {
   useEffect(() => {
     const scene = new Scene({
       id: 'map',
-    renderer: process.env.renderer,
+      renderer: process.env.renderer,
       map: new GaodeMap({
         pitch: 40,
         center: [3.438, 40.16797],
@@ -85,16 +85,16 @@ export default () => {
           .color('#00ff00')
           // .texture('plane')
           .shape('arc3d')
-          .size(20)
+          .size(10)
           // .active(true)
-          // .animate({
-          //   duration: 1,
-          //   interval: 0.2,
-          //   trailLength: 0.05,
-          // })
+          .animate({
+            duration: 1,
+            interval: 0.2,
+            trailLength: 1.0,
+          })
           .style({
             // textureBlend: 'replace',
-            lineTexture: true, // 开启线的贴图功能
+            // lineTexture: true, // 开启线的贴图功能
             iconStep: 10, // 设置贴图纹理的间距
             opacity: 1,
             sourceColor: '#f00',
