+++ dev-demos/features/line/linearline.md
@@ -1,2 +1,2 @@
 ### Line - linear 
-<code src="./demos/linearline.tsx"></code>
+<code src="./demos/linelinear.tsx"></code>
