+++ dev-demos/features/line/simpleLine_linear.md
@@ -1,3 +1,5 @@
+### Line - simpleLinear
+```tsx
 // @ts-ignore
 import {
   LineLayer,
@@ -9,18 +11,17 @@ import {
   // @ts-ignore
 } from '@antv/l7';
 // @ts-ignore
-import { GaodeMap,Map } from '@antv/l7-maps';
+import { Map,GaodeMap } from '@antv/l7-maps';
 import React, { useEffect } from 'react';
 
 export default () => {
   useEffect(() => {
     const scene = new Scene({
-      id: 'map',      
-    renderer: process.env.renderer,
+      id: 'map',
+      renderer: process.env.renderer,
       map: new GaodeMap({
         center: [105, 32],
         zoom: 4,
-        // pitch: 60
       }),
     });
 
@@ -38,10 +39,15 @@ export default () => {
               [
                 [99.228515625, 37.43997405227057],
                 [100.72265625, 27.994401411046148],
+                [110, 27.994401411046148],
+                [110, 25],
+                [100, 25],
               ],
               [
                 [108.544921875, 37.71859032558816],
                 [112.412109375, 32.84267363195431],
+                [115, 32.84267363195431],
+                [115, 35],
               ],
             ],
           },
@@ -55,8 +61,9 @@ export default () => {
             type: 'MultiLineString',
             coordinates: [
               [
-                [116.43,39.97],
-                [108.39,22.91],
+                [110, 30],
+                [120, 30],
+                [120, 40],
               ],
             ],
           },
@@ -64,68 +71,21 @@ export default () => {
       ],
     };
     const source = new Source(geoData);
-
-
-    // scene.on('zoom', e => console.log(e))
-
     const layer = new LineLayer({ blend: 'normal' })
       .source(source)
-      .size(2)
-      .shape('arc')
+      .size(10)
+      .shape('simple')
       .color('#f00')
       .style({
-        // lineType:'dash',
-        // thetaOffset: 'offset'
-        // segmentNumber: 10,
-        thetaOffset: 0.5,
+        opacity:1,
+        sourceColor: '#f00',
+        targetColor: '#0f0',
       });
-
- 
-      const point = new PointLayer({ blend: 'normal', zIndex: 1 })
-        .source([{
-          lng:116.43,
-          lat:39.97
-        },{
-          lng:108.39,
-          lat:22.91
-        }], {
-          parser: {
-            type: 'json',
-            x: 'lng',
-            y: 'lat',
-          },
-        })
-        .shape('circle')
-        .size(10)
-        .color('blue');
-        const point2 = new PointLayer({ blend: 'normal', zIndex: 1 })
-        .source([{
-          lng:116.43,
-          lat:39.97
-        },{
-          lng:108.39,
-          lat:22.91
-        }], {
-          parser: {
-            type: 'json',
-            x: 'lng',
-            y: 'lat',
-          },
-        })
-        .shape('circle')
-        .size(100000)
-        .color('blue')
-        .style({
-          opacity:0.5,
-          unit:'meter'
-        })
-      // scene.addLayer(point);
-      // scene.addLayer(point2);
-
 
     scene.on('loaded', () => {
       scene.addLayer(layer);
       scene.startAnimate();
+    
     });
   }, []);
   return (
@@ -138,3 +98,4 @@ export default () => {
     />
   );
 };
+```
