+++ packages/layers/src/line/models/arc.ts
@@ -12,15 +12,13 @@ import BaseModel from '../../core/BaseModel';
 import { ILineLayerStyleOptions } from '../../core/interface';
 import { LineArcTriangulation } from '../../core/triangulation';
 // arc dash line
-import arc_dash_frag from '../shaders/dash/arc_dash_frag.glsl';
-import arc_dash_vert from '../shaders/dash/arc_dash_vert.glsl';
+import arc_dash_frag from '../shaders/arc/arc_dash_frag.glsl';
+import arc_dash_vert from '../shaders/arc/arc_dash_vert.glsl';
 // arc normal line
-import arc_line_frag from '../shaders/line_arc_frag.glsl';
-import arc_line_vert from '../shaders/line_arc_vert.glsl';
+import arc_line_frag from '../shaders/arc/line_arc_frag.glsl';
+import arc_line_vert from '../shaders/arc/line_arc_vert.glsl';
 import { ShaderLocation } from '../../core/CommonStyleAttribute';
-// arc linear line
-import arc_linear_frag from '../shaders/linear/arc_linear_frag.glsl';
-import arc_linear_vert from '../shaders/linear/arc_linear_vert.glsl';
+
 const lineStyleObj: { [key: string]: number } = {
   solid: 0.0,
   dash: 1.0,
@@ -48,13 +46,13 @@ export default class ArcModel extends BaseModel {
     }
 
     // 转化渐变色
-    // let useLinearColor = 0; // 默认不生效
+    let useLinearColor = 0; // 默认不生效
     let sourceColorArr = [0, 0, 0, 0];
     let targetColorArr = [0, 0, 0, 0];
     if (sourceColor && targetColor) {
       sourceColorArr = rgb2arr(sourceColor);
       targetColorArr = rgb2arr(targetColor);
-      // useLinearColor = 1;
+      useLinearColor = 1;
     }
 
     if (this.rendererService.getDirty()) {
@@ -69,35 +67,24 @@ export default class ArcModel extends BaseModel {
       }
     }
     else{
-      if (sourceColor && targetColor) {
-        commonOptions = {
-          u_sourceColor: sourceColorArr,
-          u_targetColor: targetColorArr,
-          segmentNumber,
-          u_lineDir: forward ? 1 : -1
-        }
-      }
-      else{
-        commonOptions = {
-          u_animate: this.animateOption2Array(animateOption as IAnimateOption),
-          u_textSize: [1024, this.iconService.canvasHeight || 128],
-          segmentNumber,
-          u_lineDir: forward ? 1 : -1,
-          u_icon_step: iconStep,
-          u_line_texture: lineTexture ? 1.0 : 0.0, // 传入线的标识
-          u_textureBlend: textureBlend === 'normal' ? 0.0 : 1.0,
-          u_blur: 0.9,
-          u_line_type: lineStyleObj[lineType || 'solid'],
-          u_time: this.layer.getLayerAnimateTime(),
-          // // 纹理支持参数
-          // u_texture: this.texture, // 贴图
-  
-          // // 渐变色支持参数
-          // u_linearColor: useLinearColor,
-          // u_sourceColor: sourceColorArr,
-          // u_targetColor: targetColorArr,
-        };
-      }
+      commonOptions = {
+        u_animate: this.animateOption2Array(animateOption as IAnimateOption),
+        u_sourceColor: sourceColorArr,
+        u_targetColor: targetColorArr,
+        u_textSize: [1024, this.iconService.canvasHeight || 128],
+        segmentNumber,
+        u_lineDir: forward ? 1 : -1,
+        u_icon_step: iconStep,
+        u_line_texture: lineTexture ? 1.0 : 0.0, // 传入线的标识
+        u_textureBlend: textureBlend === 'normal' ? 0.0 : 1.0,
+        u_blur: 0.9,
+        u_line_type: lineStyleObj[lineType || 'solid'],
+        u_time: this.layer.getLayerAnimateTime(),
+        // // 纹理支持参数
+        // u_texture: this.texture, // 贴图
+        // 渐变色支持参数
+        u_linearColor: useLinearColor,
+      };
     }
     const commonBufferInfo = this.getUniformsBufferInfo(commonOptions);    
     return commonBufferInfo; 
@@ -124,8 +111,7 @@ export default class ArcModel extends BaseModel {
   }
 
   public getShaders(): { frag: string; vert: string; type: string } {
-    const { sourceColor, targetColor, lineType } =
-      this.layer.getLayerConfig() as ILineLayerStyleOptions;
+    const { lineType } = this.layer.getLayerConfig() as ILineLayerStyleOptions;
     if (lineType === 'dash') {
       return {
         frag: arc_dash_frag,
@@ -133,21 +119,11 @@ export default class ArcModel extends BaseModel {
         type: 'Dash',
       };
     }
-
-    if (sourceColor && targetColor) {
-      // 分离 linear 功能
-      return {
-        frag: arc_linear_frag,
-        vert: arc_linear_vert,
-        type: 'Linear',
-      };
-    } else {
-      return {
-        frag: arc_line_frag,
-        vert: arc_line_vert,
-        type: '',
-      };
-    }
+    return {
+      frag: arc_line_frag,
+      vert: arc_line_vert,
+      type: '',
+    };
   }
 
   public async buildModels(): Promise<IModel[]> {
