--- packages/layers/src/line/models/earthArc_3d.ts
@@ -1,5 +1,0 @@
-import Arc3DModel from './arc_3d';
-
-export default class EarthArc3DModel extends Arc3DModel {
- // 与arc_3d完全重合,这里暂时保留,后续看情况是否删除
-}
