+++ packages/layers/src/line/models/index.ts
@@ -1,10 +1,8 @@
 import ArcModel from './arc';
 import Arc3DModel from './arc_3d';
-import EarthArc3DModel from './earthArc_3d';
 import FlowLineModel from './flow';
 import GreatCircleModel from './great_circle';
 import LineModel from './line';
-import LinearLine from './linearline';
 import SimpleLineModel from './simpleLine';
 import LineWallModel from './wall';
 
@@ -16,8 +14,7 @@ export type LineModelType =
   | 'wall'
   | 'simple'
   | 'line'
-  | 'linearline'
-  | 'earthArc3d';
+  | 'earthArc3d'
 
 const LineModels: { [key in LineModelType]: any } = {
   arc: ArcModel,
@@ -26,9 +23,8 @@ const LineModels: { [key in LineModelType]: any } = {
   wall: LineWallModel,
   line: LineModel,
   simple: SimpleLineModel,
-  linearline: LinearLine,
-  earthArc3d: EarthArc3DModel,
   flowline: FlowLineModel,
+  earthArc3d: Arc3DModel,
 };
 
 export default LineModels;
