+++ packages/layers/src/line/models/simpleLine.ts
@@ -9,8 +9,6 @@ import BaseModel from '../../core/BaseModel';
 import { ILineLayerStyleOptions } from '../../core/interface';
 import { SimpleLineTriangulation } from '../../core/triangulation';
 import simple_line_frag from '../shaders/simple/simpleline_frag.glsl';
-// linear simple line shader
-import simle_linear_frag from '../shaders/simple/simpleline_linear_frag.glsl';
 import simple_line_vert from '../shaders/simple/simpleline_vert.glsl';
 import { ShaderLocation } from '../../core/CommonStyleAttribute';
 const { isNumber } = lodashUtil;
@@ -39,6 +37,7 @@ export default class SimpleLineModel extends BaseModel {
       u_targetColor: targetColorArr,
       // 顶点高度 scale
       u_vertexScale: vertexHeightScale,
+      useLinearColor
     };
     const commonBufferInfo = this.getUniformsBufferInfo(commonOptions);    
     return commonBufferInfo;
@@ -51,20 +50,11 @@ export default class SimpleLineModel extends BaseModel {
   public getShaders(): { frag: string; vert: string; type: string } {
     const { sourceColor, targetColor } =
       this.layer.getLayerConfig() as ILineLayerStyleOptions;
-    if (sourceColor && targetColor) {
-      // 分离 linear 功能
-      return {
-        frag: simle_linear_frag,
-        vert: simple_line_vert,
-        type: 'lineSimpleLinear',
-      };
-    } else {
       return {
         frag: simple_line_frag,
         vert: simple_line_vert,
         type: 'lineSimpleNormal',
       };
-    }
   }
 
   public async buildModels(): Promise<IModel[]> {
