+++ packages/layers/src/line/shaders/arc/arc_dash_frag.glsl
