--- packages/layers/src/line/shaders/line_arc2d_vert.glsl
@@ -1,114 +1,0 @@
-#define LineTypeSolid 0.0
-#define LineTypeDash 1.0
-#define Animate 0.0
-attribute vec4 a_Color;
-attribute vec3 a_Position;
-attribute vec4 a_Instance;
-attribute float a_Size;
-uniform mat4 u_ModelMatrix;
-uniform float segmentNumber;
-uniform vec4 u_animate: [ 1., 2., 1.0, 0.2 ];
-varying vec4 v_color;
-varying vec2 v_normal;
-
-varying float v_distance_ratio;
-uniform float u_line_type: 0.0;
-uniform vec4 u_dash_array: [10.0, 5., 0, 0];
-varying vec4 v_dash_array;
-
-#pragma include "projection"
-#pragma include "project"
-#pragma include "picking"
-
-float maps (float value, float start1, float stop1, float start2, float stop2) {
-    return start2 + (stop2 - start2) * ((value - start1) / (stop1 - start1));
-}
-
-float getSegmentRatio(float index) {
-    return smoothstep(0.0, 1.0, index / (segmentNumber - 1.));
-}
-
-float paraboloid(vec2 source, vec2 target, float ratio) {
-    vec2 x = mix(source, target, ratio);
-    vec2 center = mix(source, target, 0.5);
-    float dSourceCenter = distance(source, center);
-    float dXCenter = distance(x, center);
-    return (dSourceCenter + dXCenter) * (dSourceCenter - dXCenter);
-}
-
-vec3 getPos(vec2 source, vec2 target, float segmentRatio) {
-     float vertex_height = paraboloid(source, target, segmentRatio);
-
-    return vec3(
-    mix(source, target, segmentRatio),
-    sqrt(max(0.0, vertex_height))
-    );
-}
-vec2 getExtrusionOffset(vec2 line_clipspace, float offset_direction) {
-  // normalized direction of the line
-  vec2 dir_screenspace = normalize(line_clipspace);
-  // rotate by 90 degrees
-   dir_screenspace = vec2(-dir_screenspace.y, dir_screenspace.x);
-  vec2 offset = dir_screenspace * offset_direction * setPickingSize(a_Size) / 2.0;
-  return offset * vec2(1.0, -1.0);
-}
-vec2 getNormal(vec2 line_clipspace, float offset_direction) {
-  // normalized direction of the line
-  vec2 dir_screenspace = normalize(line_clipspace);
-  // rotate by 90 degrees
-   dir_screenspace = vec2(-dir_screenspace.y, dir_screenspace.x);
-   return reverse_offset_normal(vec3(dir_screenspace,1.0)).xy * sign(offset_direction);
-}
-float getAngularDist (vec2 source, vec2 target) {
-  vec2 delta = source - target;
-  vec2 sin_half_delta = sin(delta / 2.0);
-  float a =
-    sin_half_delta.y * sin_half_delta.y +
-    cos(source.y) * cos(target.y) *
-    sin_half_delta.x * sin_half_delta.x;
-  return 2.0 * atan(sqrt(a), sqrt(1.0 - a));
-}
-vec2 interpolate (vec2 source, vec2 target, float angularDist, float t) {
-  // if the angularDist is PI, linear interpolation is applied. otherwise, use spherical interpolation
-  if(abs(angularDist - PI) < 0.001) {
-    return (1.0 - t) * source + t * target;
-  }
-  float a = sin((1.0 - t) * angularDist) / sin(angularDist);
-  float b = sin(t * angularDist) / sin(angularDist);
-  vec2 sin_source = sin(source);
-  vec2 cos_source = cos(source);
-  vec2 sin_target = sin(target);
-  vec2 cos_target = cos(target);
-  float x = a * cos_source.y * cos_source.x + b * cos_target.y * cos_target.x;
-  float y = a * cos_source.y * sin_source.x + b * cos_target.y * sin_target.x;
-  float z = a * sin_source.y + b * sin_target.y;
-  return vec2(atan(y, x), atan(z, sqrt(x * x + y * y)));
-}
-
-void main() {
-  v_color = a_Color;
-  vec2 source = radians(a_Instance.rg);
-  vec2 target = radians(a_Instance.ba);
-  float angularDist = getAngularDist(source, target);
-  float segmentIndex = a_Position.x;
-  float segmentRatio = getSegmentRatio(segmentIndex);
-  float indexDir = mix(-1.0, 1.0, step(segmentIndex, 0.0));
-  if(u_line_type == LineTypeDash) {
-    v_distance_ratio = segmentIndex / segmentNumber;
-    float total_Distance = pixelDistance(a_Instance.rg, a_Instance.ba);
-    v_dash_array = pow(2.0, 20.0 - u_Zoom) * u_dash_array / (total_Distance / segmentNumber * segmentIndex);
-  }
-  if(u_animate.x == Animate) {
-      v_distance_ratio = segmentIndex / segmentNumber;
-  }
-  float nextSegmentRatio = getSegmentRatio(segmentIndex + indexDir);
-  v_distance_ratio = segmentIndex / segmentNumber;
-  vec4 curr = project_position(vec4(degrees(interpolate(source, target, angularDist, segmentRatio)), 0.0, 1.0));
-  vec4 next = project_position(vec4(degrees(interpolate(source, target, angularDist, nextSegmentRatio)), 0.0, 1.0));
-  v_normal = getNormal((next.xy - curr.xy) * indexDir, a_Position.y);
-  vec2 offset = project_pixel(getExtrusionOffset((next.xy - curr.xy) * indexDir, a_Position.y));
-  //  vec4 project_pos = project_position(vec4(curr.xy, 0, 1.0));
-  gl_Position = project_common_position_to_clipspace_v2(vec4(curr.xy + offset, curr.z, 1.0));
-  setPickingColor(a_PickingColor);
-}
-
